components/server/requirements.txt
Django>=3.0.0,<3.1.0
flake8>=3.7.0,<3.8.0
psycopg2>=2.8.0,<2.9.0
djangorestframework
markdown
django-filter
